build support for js and builder refactor into sep files
